app/pyproject.toml
[tool.poetry]
name = "recruitment-task"
version = "0.1.0"
description = ""
authors = ["Mateusz Ostalecki"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.2.5"
psycopg2 = "^2.9.1"
django-crispy-forms = "^1.12.0"
django-filter = "^2.4.0"
djangorestframework = "^3.12.4"
djangorestframework-camel-case = "^1.2.0"
requests = "^2.25.1"
pytest = "^6.2.4"
pytest-django = "^4.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
